Make new filte for GenresList
